Reworded heading name to reduce admin confusion
Discussed with Courtney. Measurements info box states that "It is recommended to record weight, height and head circumference for all patients." Since this is a static description that does not change when items are set as "Recommended" by the admin, we decided that that the administration field name should be changed from "Recommended" to "Fixed".