PT-2507: added a link crediting the database being used in the information pop-up.
PT-2507: clicking on the database link no longer selects the parent element.

PT-2507: hovering over the database website link displays a citation.

PT-2507: added an information widget to each database name in the ontology index, that displays a tooltip with citation information on click.

PT-2507: hovering over the database website link displays a citation.

PT-2507: added an information widget to each database name in the ontology index, that displays a tooltip with citation information on click.

PT-2507: removed the database website links from the properties file, and created api calls for this.

PT-2507: replaced the title for the help buttons on the ontology index page with something more relevant to the context.

PT-2507: implemented suggested modifications.

PT-2507: fixed javadocs.