PT-2896: Switching to another language makes name fields hidden in the pedigree node form
Fixed. Made the code more generic, and as a positive "side effect" now other fields (date of birth and death, sex, gestational age, phenotypes, disorders) will be disabled in the pedigree editor menu if they are disabled if the PhenoTips patient form.  Genes cannot be disabled yet because of an issue the `RecordConfiguration` implementation (`getAllFieldNames()` and `getEnabledFieldNames()` only return fields from the `PatientClass`, and genes are stored separately) .