PT-3853: Fixed possible bug that shows warning when there's no sensitive info
If for some reason the FamilyClass object is not initialized and its properties are NULL, the int value of the `warning` property will be '-1'. In that case, a warning will pop up, and it shouldn't. This scenario is very unlikely in our case because family records are created from the FamilyTemplate, where properties are not null, ibut t's better to do the check correctly anyway.