PT-3890 Deleted nodes are not fully removed and block drag actions
This bug happens when deleting a lone child.

The deletion logic for an only child creates an invisible placeholder node (as the algorithms require a child). The blocking element was a highlight box that is created for all nodes. This highlight box is used to indicate which nodes contain a certain gene or phenotype upon hover-over.  Since it should never appear or be used on an invisible placeholder node, it should be permissible to remove it during creation.

Tested briefly using dev tools. Need to test more thoroughly on an instance.

PT-3890 Deleted nodes are not fully removed and block drag actions

Overrode `setHighlightBox` as per code review comment. This was a better place for the code.
Tested briefly locally.

Following some advice, squashed these two commits since the second one undoes the first one.